krimiplzen/requirements.txt
django
psycopg2
channels
asgi_redis
celery
requests
rq
sorl-thumbnail
django-htmlmin
django-cachalot
django-redis
django-sirtrevor
django-meta
django-gulp
python-cloudflare-v4
django-daterange-filter
Pillow
raven
